refactor(protocol): split protocol.rs into protocol/ module
Purely mechanical refactor — zero logic changes. Extracts the
1,996-line src/protocol.rs into a protocol/ folder module with
submodules split by concern: ui (prompts/progress/log), store
(key-value state), exec (subprocess execution), metadata (project
info), and detect (project/git context detection).

Public API contract (PluginContext, OutboundMessage,
run_protocol_plugin) preserved unchanged for consumers in plugin.rs.
Updates plugin-protocol.spec.md to reference the new file layout.
